commands: move chmod code to chmod.rs
